refactor(statusline): move the statusline implementation into an own struct
