+++ helix-term/src/ui/mod.rs
@@ -8,6 +8,7 @@ mod picker;
 mod popup;
 mod prompt;
 mod spinner;
+mod statusline;
 mod text;
 
 pub use completion::Completion;
@@ -18,6 +19,7 @@ pub use picker::{FileLocation, FilePicker, Picker};
 pub use popup::Popup;
 pub use prompt::{Prompt, PromptEvent};
 pub use spinner::{ProgressSpinners, Spinner};
+pub use statusline::StatusLine;
 pub use text::Text;
 
 use helix_core::regex::Regex;
